moved alarm clock into its own folder, hope that's ok
